pyproject.toml
[tool.poetry]
name = "tz-factory"
version = "0.1.0"
description = ""
authors = ["Ayta Dzhivanov <kurt.alive.1994@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
psycopg2 = "^2.9.7"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
